+++ content_settings/settings.py
@@ -31,7 +31,7 @@ CACHE_TIMEOUT = get_setting("CACHE_TIMEOUT", 60 * 60 * 24)
 
 CACHE_SPLITER = get_setting("CACHE_SPLITER", "::")
 
-VALUES_ONLY_FROM_DB = get_setting("VALUES_ONLY_FROM_DB", False)
+VALUES_ONLY_FROM_DB = get_setting("VALUES_ONLY_FROM_DB", False) and not settings.DEBUG
 
 CONTEXT = get_setting("CONTEXT", {})